backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
psycopg2-binary
Werkzeug
